+++ CHANGELOG.md
@@ -6,7 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## Unreleased
+### Changed
 - Set javascript mode for floor panel web view to unrestricted so feature toggles will work [#5295](https://github.com/rokwire/illinois-app/issues/5295).
+- Prefer inNotOffline Connectivity's API than isOnline.
 
 ## [7.2.15] - 2025-07-24
 ### Changed